Use the --release flag to set Java version in JDK9+.
The --release flag is superior to -source and -target because --release
also compiles the Java code against the Java standard library of the
chosen version, which prevents ABI incompatibilities such as using
methods which do not exist and covariant return types changing between
Java versions. See https://openjdk.java.net/jeps/247 for more details.

The flag was added in JDK9, so we use the old flags for JDK8.